requirements.txt
certifi==2018.10.15
chardet==3.0.4
Click==7.0
Flask==1.0.2
googlemaps==3.0.2
gunicorn==19.9.0
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
psycopg2==2.7.6.1
python-dotenv==0.10.0
requests==2.20.1
urllib3==1.24.1
Werkzeug==0.14.1
